+++ src/config/config.py
@@ -26,3 +26,4 @@ class Config:
         self.autoEatFoodInInventory = True
         self.removeDeadEntities = True
         self.showMiniMap = True
+        self.cameraFollowPlayer = False